pyproject.toml
[tool.poetry]
name = "qt5_proj"
version = "0.1.0"
description = ""
authors = ["Lex Chen <clang2you@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyQt5-tools = "^5.13.2"
pyinstaller = "^3.6"
pywin32-ctypes = "^0.2.0"
pefile = "^2019.4.18"
matplotlib = "^3.1.3"
pymysql = "^0.9.3"
python-dateutil = "^2.8.1"
pandas = "^1.0.1"
pyzmq = "^19.0.0"
xlwt = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
